consolidated /programs routes to /program and linted
