--- backend/src/routes/programs.ts
@@ -1,13 +1,0 @@
-/**
- * Program route requests.
- */
-
-import express from "express";
-
-import * as ProgramsController from "../../src/controllers/programs";
-
-const router = express.Router();
-
-router.get("/", ProgramsController.getAllPrograms);
-
-export default router;
